For primitive rows fall back to more efficient converter, code reorg
